requirements.txt
jmespath
pyhcl
